also move dense_cell_range_function
