+++ eval/src/vespa/eval/instruction/dense_lambda_peek_optimizer.cpp
@@ -2,9 +2,9 @@
 
 #include "dense_lambda_peek_optimizer.h"
 #include "dense_lambda_peek_function.h"
+#include "dense_cell_range_function.h"
 #include <vespa/eval/tensor/dense/dense_tensor_view.h>
 #include <vespa/eval/tensor/dense/dense_replace_type_function.h>
-#include <vespa/eval/tensor/dense/dense_cell_range_function.h>
 #include <vespa/eval/eval/value.h>
 #include <vespa/eval/eval/node_tools.h>
 #include <vespa/eval/eval/basic_nodes.h>
